requirements.txt
astunparse==1.6.3
attrs==23.1.0
basepage==1.7.4
certifi==2023.7.22
charset-normalizer==3.2.0
docutils==0.20.1
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
MarkupSafe==2.1.3
outcome==1.2.0
packaging==23.1
pages==0.3
piglet-templates==1.3.0
pluggy==1.2.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.0
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
utils==1.0.1
wsproto==1.2.0
zipp==3.16.2

python-dotenv~=1.0.0
behave~=1.2.6
PyYAML~=6.0.1
ddt~=1.6.0
pytest-bdd~=6.1.1